CHORE: move NS API calls to ns_client.py
