+++ components/bot.py
@@ -1,16 +1,14 @@
 import logging
 from datetime import datetime, timedelta, timezone
-from typing import List, Optional
+from typing import Optional
 
-import aiohttp
 import aiomysql
 import discord
-from bs4 import BeautifulSoup as bs
 from discord.ext import commands
 
-from components.config.config_manager import configInstance
 from components.database import Database
-from components.errors import LastRecruitmentTooRecent, NotRegistered, TooManyRequests
+from components.errors import LastRecruitmentTooRecent, NotRegistered
+from components.ns_client import NSClient
 from components.queue import QueueManager
 from components.recruiter import Recruiter
 
@@ -18,14 +16,6 @@ logger = logging.getLogger("main")
 
 
 class Bot(commands.Bot):
-    @property
-    def headers(self) -> dict:
-        return self._headers
-
-    @property
-    def session(self) -> aiohttp.ClientSession:
-        return self._session
-
     @property
     def pool(self) -> aiomysql.Pool:
         return self._pool
@@ -35,49 +25,22 @@ class Bot(commands.Bot):
         return self._db
 
     @property
-    def ratelimit(self) -> Optional[int]:
-        """The number of requests that NS will accept in within one bucket"""
-        return self._ratelimit
-
-    @property
-    def policy(self) -> Optional[int]:
-        """The length of an NS API bucket in seconds"""
-        return self._policy
-
-    @property
-    def remaining(self) -> Optional[int]:
-        """The number of requests remaining in the current NS API bucket"""
-        return self._remaining
-
-    @property
-    def reset_in(self) -> Optional[int]:
-        """The number of seconds until the current NS API bucket is reset"""
-        return self._reset_in
-
-    @property
-    def request_timestamps(self) -> List[datetime]:
-        """A list of timestamps for each request made in the current NS API bucket"""
-        return self._request_timestamps
+    def ns(self) -> NSClient:
+        return self._ns
 
     @property
     def queue_manager(self) -> QueueManager:
         """The recruitment queue"""
         return self._queue_list
 
-    def __init__(self, session: aiohttp.ClientSession, ql: QueueManager, pool: aiomysql.Pool):
+    def __init__(self, ns: NSClient, ql: QueueManager, pool: aiomysql.Pool):
         intents = discord.Intents.default()
 
         super().__init__(command_prefix="!", intents=intents)
 
-        self._headers = {"User-Agent": f"Aperta Recruitment Bot, developed by nation=upc, run by {configInstance.data.operator}"}
-        self._session = session
         self._pool = pool
         self._db = Database(pool)
-        self._ratelimit = None
-        self._policy = None
-        self._remaining = None
-        self._reset_in = None
-        self._request_timestamps = []
+        self._ns = ns
         self._queue_list = ql
 
     async def setup_hook(self):
@@ -122,34 +85,6 @@ class Bot(commands.Bot):
         )
 
         return row[0]
-
-    async def request(self, url: str) -> bs:
-        current_time = datetime.now(timezone.utc)
-
-        if self._ratelimit:
-            # Check if we are exceeding the user defined maximum number of requests in a bucket
-            while len(self._request_timestamps) >= configInstance.data.period_max:
-                elapsed = (current_time - self._request_timestamps[0]).total_seconds()
-
-                if elapsed > self._policy:
-                    del self._request_timestamps[0]
-                else:
-                    raise TooManyRequests(self._policy - elapsed)
-
-            # Check if we are exceeding the NS defined maximum number of requests in a bucket
-            if self._remaining and self._remaining <= 1:
-                raise TooManyRequests(self.reset_in if self.reset_in else 30)
-
-        async with self._session.get(url, headers=self._headers) as resp:
-            self._ratelimit = int(resp.headers["RateLimit-limit"])
-            self._policy = int(resp.headers["RateLimit-policy"].split(";w=")[1])
-            self._remaining = int(resp.headers["RateLimit-remaining"])
-            self._reset_in = int(resp.headers["RateLimit-reset"])
-            self._request_timestamps.append(current_time)
-
-            text = await resp.text()
-
-            return bs(text, "xml")
 
     async def get_recruiter_id(self, user: discord.User, channel_id: int) -> Optional[int]:
         row = await self._db.fetch_one(
